+++ android/app/src/main/kotlin/com/jabook/app/jabook/widget/PlayerWidgetProvider.kt
@@ -144,7 +144,7 @@ public class PlayerWidgetProvider : AppWidgetProvider() {
                     // Wait for controller with timeout (faster for widget UX)
                     controller =
                         controllerFuture?.get(
-                            com.jabook.app.jabook.audio.MediaControllerConstants.WIDGET_TIMEOUT_SECONDS,
+                            com.jabook.app.jabook.audio.MediaControllerConstants.WIDGET_TIMEOUT_SECONDS.toLong(),
                             TimeUnit.SECONDS,
                         )
